M1 fixes: confidence->real, cefr-map SSOT, llm.ts JSON.parse guard
Applies code-review findings on the M1 diff:
- intake_batch_answers.confidence: integer -> real to match Zod 0..1
- lib/cefr-map.ts: drop duplicated certKinds; import from schemas/_shared
- lib/llm.ts: wrap JSON.parse with schemaName + tier context
- students.local_salary_usd: Zod .int() to match DB
- students.email: UNIQUE
- roadmap.md: RESEND_FROM -> RESEND_FROM_EMAIL

+++ apps/web/schemas/student.ts
@@ -7,7 +7,7 @@ export const studentSchema = z.object({
   email: z.email(),
   name: z.string().min(1),
   whatsapp: z.string().min(5),
-  local_salary_usd: z.number().positive().optional(),
+  local_salary_usd: z.number().int().positive().optional(),
   consent_public_profile: z.boolean(),
   public_slug: z.string().min(3),
   onboarded_at: z.iso.datetime().optional(),
